Switch from u64 to usize

+++ src/lib.rs
@@ -1,7 +1,7 @@
-pub fn get_halt_arg() -> u64 {
+pub fn get_halt_arg() -> usize {
     std::env::args()
         .nth(1)
         .expect("Provide a limit.")
         .parse::<f64>()
-        .expect("Failed to parse limit") as u64
+        .expect("Failed to parse limit") as usize
 }
